+++ common/src/main/java/name/dashkal/minecraft/hexresearch/interop/impl/HexGloopInteropImpl.java
@@ -29,7 +29,7 @@ public class HexGloopInteropImpl {
     }
 
     /**
-     * Offers a mind to HexGloop (IDynamicTarget) to see if it wants to accept the mind.
+     * Offers a mind to HexGloop (IDynamicFlayTarget) to see if it wants to accept the mind.
      *
      * @param ctx the casting context in which we are imbuing
      * @param blockPos block position of the target block